Use Node's url module
I originally wanted to support relative paths to sourcemaps, but at the
time I wrote this, I didn't have an example handy that was generated
like this. However, thanks to @digitarald, I have an example to test
against.

  https://perf-html.io/9eea8cd330febdcdf53d.bundle.js

To make this work, I decided to start using Node's url module, which has
a `URL` class on it that can take an input URL and a base URL.

  https://nodejs.org/api/url.html#url_constructor_new_url_input_base

Fixes #1