Improve sourceMappingURL finding
I received a report that the sourceMappingURL was not found in some
files. I noticed that this was because there was a trailing newline in
these files, but this function was written in a way that did not account
for that. In this commit I attempt to make that better by instead of
just looking at the last line, using a regex to search through the last
5000 characters.